Renamed/Refactored TOC and Title code
